+++ dev_history.R
@@ -2,6 +2,8 @@
 usethis::use_build_ignore("dev_history.R")
 usethis::use_build_ignore("reference/")
 usethis::use_build_ignore("_pkgdown.yml")
+usethis::use_build_ignore("gitdown")
+usethis::use_git_ignore("gitdown")
 
 usethis::use_git(message = "init commitdown")
 usethis::use_mit_license("ThinkR")