+++ NEWS.md
@@ -4,6 +4,7 @@ Breaking changes
 * `names_section` parameter is replaced by a named vector in `pattern`
 
 Other changes
+* List all patterns of `pattern.table` even if no commits associated
 * Add parameter `pattern.table` to allow for user specific pattern titles
 * `create_vignette_last_modif()` creates a vignette with dates of files modification in git
 * Added summary tables for each section